Squash bug that was passing in bad `update-file-btn` data to searchForFile
-- Discovered that inside `onUpdateFile`, we were assigning
searchObject.id to the first instance of the `.update-file-btn` class's
id, which was always the first resource in the table. This was passing
that first resource into our searchForFile function, and always returning
that object.
-- If we use `this`, we can select specifically the button we clicked on
(thinking this is `event.target`), and access that data attribute
-- Now, the proper `file-id` is assigned to searchObject.id, and we
can accurately search for and return the proper element

Paired with @nandodonaire, @skinnybuff, and @TedCart